Phase 6: Export 파이프라인 완성 — YUV 직접 전달 + AAC 오디오 인코딩
- YUV420P 직접 파이프라인: Decoder→YUV420P→Encoder (RGBA 이중 변환 제거, 무손실 색공간)
- H.264 비디오 인코딩: libx264 CRF / 시스템 인코더 bitrate 자동 선택
- AAC 오디오 인코딩: 48kHz stereo 192kbps, f32→FLTP deinterleave
- AudioMixer Export 통합: 다중 클립 합성 + 소프트 클리핑
- decode_range f64 정밀도: i64 truncation으로 인한 30Hz 클릭 노이즈 수정
- 비ASCII(한글) 출력 경로: TEMP 임시파일 → 최종 경로 이동
- LANCZOS 디코더 스케일러: Export용 고품질 리사이즈
- Fallback bitrate 수정: non-libx264 인코더 적정 bitrate 계산
- README 업데이트: Phase 6 반영, 기술 스택 갱신

+++ rust-engine/src/encoding/audio_mixer.rs
@@ -71,8 +71,8 @@ impl AudioMixer {
                 None => continue,
             };
 
-            // PCM 디코딩
-            let samples = match decoder.decode_range(source_start, duration_ms as i64) {
+            // PCM 디코딩 (duration_ms를 f64로 전달 — i64 truncation하면 매 프레임 샘플 부족 → 노이즈)
+            let samples = match decoder.decode_range(source_start, duration_ms) {
                 Ok(s) => s,
                 Err(e) => {
                     eprintln!("[AUDIO_MIX] 디코딩 실패 {}: {}", file_path, e);
